+++ src/components/App.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
-import BooksList from './BookList';
-import BooksForm from './BooksForm';
+import BooksList from '../containers/BookList';
+import BooksForm from '../containers/BooksForm';
 
 const App = ({ books }) => (
   <div>
